requirements.txt
psycopg2-binary==2.9.11
pgvector==0.4.2
sentence-transformers==5.3.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
transformers==5.3.0
numpy==2.4.3
fastapi==0.115.0
uvicorn==0.30.0
scikit-learn==1.6.1
openai==1.60.0
